changed hl_uint32 from unsigned long to unsigned int to fix sha256 issues on 64Bit systems
git-svn-id: http://svn.code.sf.net/p/hashlib2plus/code@61 9fabf1d2-8a3a-0410-866c-cb411183c2a1

+++ trunk/src/hl_types.h
@@ -54,7 +54,7 @@ typedef unsigned short int 	hl_uint16;
 /**
  * at least 4 Byte
  */
-typedef unsigned long int hl_uint32;
+typedef unsigned int hl_uint32;
 
 /**
  * at least 8 Byte
